requirements.txt
attrs==22.2.0
colorama==0.4.6
iniconfig==2.0.0
numpy==1.24.2
packaging==23.0
pluggy==1.0.0
pytest==7.2.2
pyyaml==6.0
six==1.16.0
